requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.80
botocore==1.27.80
branca==0.5.0
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
django-crispy-forms==1.14.0
django-storages==1.13.1
folium==0.12.1.post1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.2
urllib3==1.26.12
